# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EcoJharkhand AI tourism platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Destination {
  id: number;
  name: string;
  /** waterfall, forest, village, temple, eco */
  category: string;
  description: string;
  location: string;
  /** Eco friendliness score 1-100 */
  ecoScore: number;
  tribalStory: string;
  culturalInfo: string;
  traditions: string;
  imageUrl: string;
  isHiddenGem: boolean;
  rating: number;
  visitCount: number;
  /** @nullable */
  bestSeason?: string | null;
  /** @nullable */
  entryFee?: number | null;
}

export interface DestinationInput {
  name: string;
  category: string;
  description: string;
  location: string;
  ecoScore?: number;
  tribalStory: string;
  culturalInfo: string;
  traditions: string;
  imageUrl: string;
  isHiddenGem?: boolean;
  bestSeason?: string;
  entryFee?: number;
}

export type DestinationStatsCategoryBreakdownItem = {
  category: string;
  count: number;
};

export interface DestinationStats {
  categoryBreakdown: DestinationStatsCategoryBreakdownItem[];
  topRated: Destination[];
  totalCount: number;
  hiddenGemsCount: number;
}

export interface ItineraryRequest {
  /**
     * @minimum 1
     * @maximum 14
     */
  days: number;
  /** Budget in INR */
  budget: number;
  /** e.g. ["nature", "tribal culture", "waterfalls", "adventure"] */
  interests: string[];
  /** solo, couple, family, group */
  travelType: string;
  startLocation?: string;
}

export interface ItineraryDay {
  day: number;
  destinations: string[];
  activities: string[];
  accommodation: string;
  estimatedCost: number;
  ecoScore: number;
}

export interface Itinerary {
  days: ItineraryDay[];
  totalEstimatedCost: number;
  ecoTips: string[];
  summary: string;
}

export interface EcoScoreEntry {
  id: number;
  userName: string;
  action: string;
  points: number;
  createdAt: string;
}

export interface EcoUser {
  name: string;
  totalPoints: number;
  /** seed, sapling, tree, forest, guardian */
  badge: string;
  rank: number;
  actions: EcoScoreEntry[];
}

export interface EcoScoreInput {
  userName: string;
  /** eco_transport, homestay, local_artisan, no_plastic, trail_cleanup */
  action: string;
  points: number;
}

export interface Review {
  id: number;
  userName: string;
  destinationId: number;
  destinationName: string;
  text: string;
  rating: number;
  /** positive, neutral, negative */
  sentiment: string;
  createdAt: string;
}

export interface ReviewInput {
  userName: string;
  destinationId: number;
  text: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
}

export interface SentimentSummary {
  positive: number;
  neutral: number;
  negative: number;
  total: number;
  recentReviews: Review[];
  averageRating?: number;
}

export interface Handicraft {
  id: number;
  name: string;
  artisan: string;
  village: string;
  /** pottery, weaving, painting, jewelry, bamboo, wood */
  category: string;
  price: number;
  description: string;
  imageUrl: string;
  inStock: boolean;
}

export interface BudgetRequest {
  days: number;
  /** solo, couple, family, group */
  travelType: string;
  /** budget, homestay, mid-range, luxury */
  accommodationType: string;
  /** public, shared, private, eco */
  transportType: string;
  destinations?: string[];
  includeLocalShopping?: boolean;
}

export type BudgetResultBreakdownItem = {
  category: string;
  amount: number;
  percentage: number;
};

export interface BudgetResult {
  transport: number;
  accommodation: number;
  food: number;
  activities: number;
  localShopping: number;
  total: number;
  perPersonCost: number;
  breakdown: BudgetResultBreakdownItem[];
}

export type DashboardStatsCategoryBreakdownItem = {
  category: string;
  count: number;
};

export type DashboardStatsRecentActivityItem = {
  type: string;
  message: string;
  createdAt: string;
};

export interface DashboardStats {
  totalDestinations: number;
  totalReviews: number;
  totalEcoUsers: number;
  totalHandicrafts: number;
  sentimentBreakdown: SentimentSummary;
  categoryBreakdown: DashboardStatsCategoryBreakdownItem[];
  topEcoUsers: EcoUser[];
  recentActivity: DashboardStatsRecentActivityItem[];
}

export type ListDestinationsParams = {
/**
 * Filter by category (waterfall, forest, village, temple, eco)
 */
category?: string;
/**
 * Search by name or description
 */
search?: string;
/**
 * Filter featured/hidden gems
 */
featured?: boolean;
};

export type ListReviewsParams = {
destinationId?: number;
/**
 * Filter by sentiment (positive, neutral, negative)
 */
sentiment?: string;
};

export type ListHandicraftsParams = {
category?: string;
search?: string;
};