# lib/api-zod/src/generated/types/budgetRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EcoJharkhand AI tourism platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BudgetRequest {
  days: number;
  /** solo, couple, family, group */
  travelType: string;
  /** budget, homestay, mid-range, luxury */
  accommodationType: string;
  /** public, shared, private, eco */
  transportType: string;
  destinations?: string[];
  includeLocalShopping?: boolean;
}