# lib/api-zod/src/generated/types/destination.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EcoJharkhand AI tourism platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Destination {
  id: number;
  name: string;
  /** waterfall, forest, village, temple, eco */
  category: string;
  description: string;
  location: string;
  /** Eco friendliness score 1-100 */
  ecoScore: number;
  tribalStory: string;
  culturalInfo: string;
  traditions: string;
  imageUrl: string;
  isHiddenGem: boolean;
  rating: number;
  visitCount: number;
  /** @nullable */
  bestSeason?: string | null;
  /** @nullable */
  entryFee?: number | null;
}