# lib/api-zod/src/generated/types/destinationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EcoJharkhand AI tourism platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DestinationInput {
  name: string;
  category: string;
  description: string;
  location: string;
  ecoScore?: number;
  tribalStory: string;
  culturalInfo: string;
  traditions: string;
  imageUrl: string;
  isHiddenGem?: boolean;
  bestSeason?: string;
  entryFee?: number;
}