# lib/api-zod/src/generated/types/destinationStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EcoJharkhand AI tourism platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Destination } from './destination';
import type { DestinationStatsCategoryBreakdownItem } from './destinationStatsCategoryBreakdownItem';

export interface DestinationStats {
  categoryBreakdown: DestinationStatsCategoryBreakdownItem[];
  topRated: Destination[];
  totalCount: number;
  hiddenGemsCount: number;
}