# lib/api-zod/src/generated/types/ecoScoreEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EcoJharkhand AI tourism platform API
 * OpenAPI spec version: 0.1.0
 */

export interface EcoScoreEntry {
  id: number;
  userName: string;
  action: string;
  points: number;
  createdAt: string;
}