# lib/api-zod/src/generated/types/ecoScoreInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EcoJharkhand AI tourism platform API
 * OpenAPI spec version: 0.1.0
 */

export interface EcoScoreInput {
  userName: string;
  /** eco_transport, homestay, local_artisan, no_plastic, trail_cleanup */
  action: string;
  points: number;
}