# lib/api-zod/src/generated/types/itinerary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EcoJharkhand AI tourism platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ItineraryDay } from './itineraryDay';

export interface Itinerary {
  days: ItineraryDay[];
  totalEstimatedCost: number;
  ecoTips: string[];
  summary: string;
}