# lib/api-zod/src/generated/types/itineraryDay.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EcoJharkhand AI tourism platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ItineraryDay {
  day: number;
  destinations: string[];
  activities: string[];
  accommodation: string;
  estimatedCost: number;
  ecoScore: number;
}