# lib/api-zod/src/generated/types/sentimentSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EcoJharkhand AI tourism platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Review } from './review';

export interface SentimentSummary {
  positive: number;
  neutral: number;
  negative: number;
  total: number;
  recentReviews: Review[];
  averageRating?: number;
}